% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bus_stop_list}
\alias{bus_stop_list}
\title{Bus stop lists}
\format{A large list with 10133 elements.}
\usage{
bus_stop_list
}
\description{
A dataset containing the bus stops in Dublin
}
\keyword{dataset}
