% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_api.R
\name{dart_api}
\alias{dart_api}
\title{Call API for real time info for a DART station}
\usage{
dart_api(station_name)
}
\arguments{
\item{station_name}{A station name, short version.}
}
\value{
A data frame containing the times until the next train at the selected station.
}
\description{
Call API for real time info for a DART station
}
\examples{
\dontrun{dart_api("tara")}

}
