% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_info.R
\name{dart_info}
\alias{dart_info}
\title{Get the real time info for a DART station}
\usage{
dart_info(station_name)
}
\arguments{
\item{station_name}{A station name, short version.}
}
\value{
A data frame containing the times until the next train at the selected station.
}
\description{
Get the real time info for a DART station
}
\examples{
\dontrun{dart_stop_info("tara")}

}
