% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_scrape_info.R
\name{db_scrape_info}
\alias{db_scrape_info}
\title{Real time info for multiple bus stops}
\usage{
db_scrape_info(stop_numbers)
}
\arguments{
\item{stop_numbers}{A vector of bus stop numbers.}
}
\value{
A data frame containing the times until the next buses at the selected stops.
}
\description{
This function uses the \link{db_scrape_stop_route_info} function to collect info for multiple stops.
}
\examples{
\dontrun{db_scrape_multi_stop_info(c(334, 336))}

}
