% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_scrape_stop_route_info.R
\name{db_scrape_stop_route_info}
\alias{db_scrape_stop_route_info}
\title{Real time info for singlebus stops}
\usage{
db_scrape_stop_route_info(stop_number)
}
\arguments{
\item{stop_number}{A numeric value for the bus stop number.}
}
\value{
A data frame containing the times until the next buses at the selected stop.
}
\description{
This function uses the Dublin Bus website to retrieve the real time information about a single bus stop.
}
\examples{
\dontrun{db_scrape_stop_route_info(334)}

}
