\name{list2df}
\alias{list2df}
\title{List to Data Frame}
\description{
The function list2df converts a list to a data.frame.
}
\usage{
list2df(x)
}
\arguments{
 \item{x}{a list with each entry being a vector.}
}
\details{
This function converts a list of vectors into a two-column data.frame.
}
\value{
A two-column data.frame, with first column being the elements in \code{x} and the second column being the list entry name/id.
}
\seealso{\code{\link{data.frame}}}
\examples{
##not run###
}
