% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luas_info.R
\name{luas_info}
\alias{luas_info}
\title{API call for Luas real time info}
\usage{
luas_info(stop_name)
}
\arguments{
\item{stop_name}{Abbreviated stop name}
}
\description{
API call for Luas real time info
}
