% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luas_stops}
\alias{luas_stops}
\title{Luas stops}
\format{
A list with 167 elements.
\describe{
  \item{StopID}{Stop ID}
  \item{Name}{Name}
  \item{Abbreviation}{Stop name abbreviation}
  \item{LineID}{Line ID 1 = Red, 2 = Green}
  \item{SortOrder}{Station code}
  \item{IsEnabled}{}
  \item{IsParkAndRide}{}
  \item{IsCycleAndRide}{}
  \item{ZoneCountA}{}
  \item{ZoneCountB}{}
  \item{Latitude}{}
  \item{Longitude}{}
  \item{IrishName}{Name in Gaeilge}
  \item{Line}{Line name, red or green}
  }
}
\usage{
luas_stops
}
\description{
A dataset containing all luas stops
}
\keyword{dataset}
