% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctl.R
\docType{package}
\name{rctl}
\alias{rctl}
\alias{rctl-package}
\title{Can be used to install all rctl command line tools required packages (CRAN, bioconductor and GitHub).}
\description{
Can be used to install all rctl command line tools required packages (CRAN, bioconductor and GitHub).
}
\seealso{
Useful links:

\url{https://github.com/openanno/rctl/src/rctl}

Report bugs at \url{https://github.com/openanno/rctl/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
