% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib-length-h.R
\name{r_length}
\alias{r_length}
\alias{r_short_length}
\title{Length}
\arguments{
\item{x}{A sexp to compute the length of.}
}
\value{
For \code{r_length()}, a \code{r_ssize_t} containing the length of \code{x}.

For \code{r_short_length()}, a \code{r_short_ssize_t} containing the
short length of \code{x}.
}
\description{
\itemize{
\item \code{r_length()} computes the length of \code{x}. This returns a \code{r_ssize_t}
and supports long vectors.
\item \code{r_short_length()} computes the old style length of \code{x}. This returns a
\code{r_short_ssize_t} and does not support long vectors.
}
}
