% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_train_stations}
\alias{tidy_train_stations}
\title{Train stations (tidy)}
\format{
A list with 167 elements.
}
\usage{
tidy_train_stations
}
\description{
A dataset containing all train stations in Ireland
}
\keyword{dataset}
