% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_stations}
\alias{train_stations}
\title{Train stations}
\format{A list with 167 elements.
\describe{
  \item{StationDesc}{Full station name}
  \item{StationAlias}{Alternative name}
  \item{StationLatitude}{Latitude}
  \item{StationLongitude}{Longitude}
  \item{StationCode}{Station code}
  \item{StationId}{Stiotion ID}
  }}
\usage{
train_stations
}
\description{
A dataset containing all train stations in Ireland
}
\keyword{dataset}
